Improve test time and add a small suite
